station: have station_transition_start take a list of BSS's
This prepares for the FT-over-Air refactor, which will allow a
candidate list to be provided to netdev_fast_transition. For now
generate this candidate list but use the first (best) entry for the
roam.

This does slightly change the logic and all roam candidates are
added to stations->bss_list (where before only the best candidate
was), but this shouldn't effect how IWD behaves externally.