wiphy: only do global GET_REG once
Getting the regulatory domain for self-managed wiphys requires
providing the wiphy ID, but this is not the case for any device
under the global regulatory domain. This is not tied to any
specific wiphy so there is no need to supply a wiphy ID or do
a dump for each wiphy.

Now GET_REG will be tracked globally if its a global call and
will only be called once.