wiphy: support country code override
Depending on the device the regulatory domain may never get
set on its own. This is worked around in wpa_supplicant by
allowing the user to supply a country code. The kernel
_should_ set this on its own based on the docs, but ultimately
will fall back to user defined if the driver doesn't set it.

IWD now needs to work around this problem by allowing users
to supply the country code in main.conf.

A few notes here:
 - Setting the country may not effect the actual regulatory
   domain. This is because the kernel still remains in control.
   Any request is merely a hint to the kernel.
 - This only applies to devices under the global regulatory
   domain. Self-managed devices will not allow setting the
   regdom.