netdev: try catching buggy driver behavior to avoid hangs
It was reported that IWD hangs during roaming due to the kernel never
sending any expected events after authentication. This was likely
due to buggy firmware but ultimately got IWD into an unrecoverable
state waiting for an event which never came.

The only indication of this to userspace was a DEL_STATION event
which happens after the successful authenticate event. Now IWD will
watch for this and if the conditions are right IWD will fail the
connection.