scan: make scan requests fail if trigger returns -EBUSY
If a CMD_TRIGGER_SCAN request fails with -EBUSY, iwd currently assumes
that a scan is ongoing on the underlying wdev and will retry the same
command when that scan is complete. It gets notified of that completion
via the scan_notify() function, and kicks the scan logic to try again.

However, if there is another wdev on the same wiphy and that wdev has a
scan request in flight, the kernel will also return -EBUSY. In other
words, only one scan request per wiphy is permitted.

As an example, the brcmfmac driver can create an AP interface on the
same wiphy as the default station interface, and scans can be triggered
on that AP interface.

If -EBUSY is returned because another wdev is scanning, then iwd won't
know when it can retry the original trigger request because the relevant
netlink event will arrive on a different wdev. Indeed, if no scan
context exists for that other wdev, then scan_notify will return early
and the scan logic will stall indefinitely.

From a user perspective, this leads to a situation where iwd reports
over DBus that it is not scanning, but it never queues up another
periodic scan. Or if a scan was explicitly requested over DBus and the
trigger for that scan ended in an -EBUSY for the aforementioned reason,
the DBus method call stall indefinitely.

Since scan contexts are handled on a per-wdev basis, and since multiple
wdevs may in principle share the same wiphy, solve the above problem by
instead having the scan module simply propagate the -EBUSY error. In
practice this will mean that scans may fail more often, in which case at
least one default error message will be logged each time:

iwd[1451]: Received error during CMD_TRIGGER_SCAN: Device or resource busy (16)

It can be debated whether or not this message is needed: maybe the
upper-layer may wish to implement its own retry logic instead. In the
case of a periodic scan for example, a failure is met with ambivalence:
no additional error is logged, and a new scan is simply scheduled
according to the periodic scan interval. In other cases, a scan might be
part of a DBus request to connect to a hidden network. Here, the error
will be different:

Failure to connect because the network wasn't found after scanning:

    $ iwctl station wlan0 connect-hidden foobar
    Object not found

Failure to connect because another scan was active on the same wiphy:

    $ iw dev uap0 scan >/dev/null &
    $ iwctl station wlan0 connect-hidden foobar
    Operation failed

This is the trade-off made by this change in order not to complicate the
scan context with lookup-by-wiphy and an internal retry mechanism, while
still providing robustness and not stalling the scan logic of the
daemon.

Additionally, in the event of a failed scan trigger via the DBus scan
method, do not automatically jumpstart the autoconnect logic, as this
would not make sense.