Revert "scan: Don't callback on SCAN_ABORTED"
This reverts commit 6051a1495227bbe7ba357f6995d4dbe4a4862331.

The blamed commit argues that the periodic scan callback doesn't do
anything useful in the event of an aborted scan, but this is not
entirely true. In particular, the callback is responsible for resetting
the station scanning state to false. Failure to do so means that the
station scanning state remains true indefinitely if a periodic scan is
aborted, and no further periodic scans can be performed by iwd.