wiphy: add wiphy_check_{frequency,band}
Add two new APIs as a replacement for wiphy_get_disabled_freqs().

wiphy_check_frequency() serves as a way to check a single frequency
for certain attributes, currently: supported, disabled, and no-IR.

wiphy_check_band() similarly checks the entire band. This is mainly
targeted at station which needs to check if a given band is disabled
and is only meant to be used with the supported and disabled flags.