station: delay roam until fully connected
If DHCP/netconfig takes a long time to finish IWD could end up trying
to roam to another BSS. This could happen if the signal strength is
below the set threshold while connecting. Once authenticated to the
AP IWD would get a CQM event and start a roam timer. If DHCP did not
finish before that timer IWD would start a roam. IWD would not handle
this well since a roam doesn't trigger DHCP to start and likely the
network infrustructure would be very confused.

To fix this continue to delay the roam attempt until IWD reaches the
connected state.