test-runner: fix __str__ for namespace processes
When printing out a namespace the Process.get_all() call was
returning all processes running, even those not in the namespace
being printed. Modifying get_all() is tricky since its a class
method (and used elsewhere for ALL processes) so instead filter
the results to only show processes in the namespace being printed.