hwsim: refactor rules processing into separate file
This extends hwsim to allow for external/custom frame processing
by allowing separate modules to register frame callbacks and
expose APIs to send frames back to the kernel.

One benefit of this is the hwsim rules code can be separated out
which reduces the size of hwsim.c and isolates that functionality.
Another benefit is that additional processing modules could be
dropped in cleanly without modifying existing source code.

The goal here is to isolate the basic core functions of hwsim:
 - Create, list, destroy radios
 - Register to medium to kernel
 - Track radios and interface creation/removal