build: install iwd-decrypt-profile
If profile encryption is enabled its not out of the question that
somebody might need to decrypt it. This isn't a problem building
from source but IWD as a package likely won't include the decrypt
tool unless its specifically copied out. Add iwd-decrypt-profile
to the list of installed programs like iwctl, iwmon, etc

This also fixes a minor bug in Makefile.am which was building both
iwd-profile-decrypt and probe-req if DAEMON was true. This should
be TOOLS.