RFC: Support full profile sharing via DPP 3rd party attributes
If an IWD profile contains network-specific settings which are
required to utilize the network correctly configuring via DPP
will not carry over those settings to the enrollee. The DPP
configuration object only contains the SSID/PSK to connect and
anything else set in the configurators profile is not included.

This is likely something that the majority of users will not
need (most networks don't need additional settings) but if the
network does it would be convenient for the configurator to send
over its exact configuration to the enrollee. This is useful for
an automated use case where a configuration should be consistent
across all devices.

DPP allows for arbitrary 3rd party attributes in the configuration
object (section 4.5.2) which can be used to communicate additional
settings.

The plan is to define a new object within the overall
configuration object who's keys are IWD profile groups and values
are objects containing settings for those groups:

{
  "ssid": "my_ssid",
  ... main configuration object ...

  ... The IWD profile, converted to JSON ...
  "/net/connman/iwd": {
    "Network": {
      "MutlicastDNS": "true"
    },
    "IPv4": {
      "SendHostname": "true"
    },
    ... etc ...
  }
}

The "/net/connman/iwd" object could then be parsed by the enrollee
(potentially if the feature is enable in main.conf?) and set to
the profile as it is now with the passphrase/psk.

Several profile values don't apply here like MAC/IP address
overrides. Mainly the settings that do matter would be:

[IPv4].SendHostname
[Network].MulticastDNS
[Settings] (most values here, except AddressOverride)

The other IPv4/v6 settings likely aren't useful since they
pertain mostly to static configurations and are obtained from
DHCP. The [Security] group would be omitted entirely since its
already set by the main configuration and contains generated
values like SAE-PT-Group's.