netdev: hack back in disconnect event
Not sure how we want to address this, but FT is special in that it
has no connect_cb but can still trigger an associate timeout.
Currently if FT times out associating IWD will hang indefinitely.

This behavior can be tested with the prior autotest patches (without
this patch applied).

Fixes: 30c6a10f28 ("netdev: Separate connect_failed and disconnected paths")