scan: limit COLOCATED_6GHZ flag to 6ghz devices
It was seen that this flag seems to cause issues when in AP mode on
brcmfmac devices (e.g. the raspberry Pi 3). When in AP mode an a
scan is issued clients will disconnect. After testing this behavior
was isolated to the use of the COLOCATED_6GHZ flag.

Besides working around the problem on this specific hardware the
patch itself makes sense as a non-6GHz capable device shouldn't use
this flag anyways.

As stated in the patch comment, this isn't really a catch all
workaround since the flag is still used for devices supporting 6GHz.
If additional hardware exhibits this behavior we may need additional
changes like a hardware blacklist or an explicit option to disable
the flag.

Reported-By: Keith G <ys3al35l@gmail.com>