build: make missing rst2man non-fatal
The configure.ac script does not fail to configure if rst2man is missing
but instead sets RUN_RST2MAN depending on if its available, i.e. makes
it an optional requirement. But then in Makefile.am if RUN_RST2MAN is
not set it fails fatally and will not continue with the build.

Instead remove the false return so the build can continue. Another
option would be to put a hard requirement on rst2man and not allow
configure to proceed, but based on the existing logic it did not seem
like this was the original intent.